PR fixes 2
