+++ components/org.wso2.carbon.identity.oauth/src/main/java/org/wso2/carbon/identity/oauth2/util/OAuth2Util.java
@@ -4404,85 +4404,4 @@ public class OAuth2Util {
         return IdentityTenantUtil.getTenantDomainFromContext();
     }
 
-    /**
-     * Checks if the password had expired.
-     *
-     * @param tenantDomain        The tenant domain of the user trying to authenticate.
-     * @param tenantAwareUsername The tenant aware username of the user trying to authenticate.
-     * @return True if the password had expired.
-     * @throws IdentityOAuth2Exception if the authentication failed for the user trying to login.
-     */
-    public static boolean isUserPasswordExpired(String tenantDomain, String tenantAwareUsername)
-            throws IdentityOAuth2Exception {
-
-        org.wso2.carbon.user.api.UserStoreManager userStoreManager;
-        UserRealm userRealm;
-        try {
-            int tenantId = IdentityTenantUtil.getTenantId(tenantDomain);
-            RealmService realmService = IdentityTenantUtil.getRealmService();
-            userRealm = realmService.getTenantUserRealm(tenantId);
-            userStoreManager = userRealm.getUserStoreManager();
-        } catch (UserStoreException e) {
-            throw new IdentityOAuth2Exception("Error occurred while loading user manager from user realm", e);
-        }
-
-        String passwordLastChangedTime;
-        String lastCredentialUpdateClaimURI = PasswordPolicyConstants.LAST_CREDENTIAL_UPDATE_TIMESTAMP_CLAIM;
-        String createdClaimURI = PasswordPolicyConstants.CREATED_CLAIM;
-        try {
-            // trying to first get a value for the 'lastPasswordUpdateTime' claim and if that fails the fallback is
-            // the 'created' claim
-            passwordLastChangedTime = getClaimValue(userStoreManager,lastCredentialUpdateClaimURI, tenantAwareUsername)
-                    .orElse(convertCreatedDateToEpochString(
-                            getClaimValue(userStoreManager, createdClaimURI, tenantAwareUsername).get()));
-        } catch (UserStoreException | ParseException e) {
-            throw new IdentityOAuth2Exception("Error occurred while loading user claim", e);
-        }
-
-        long passwordChangedTime;
-        passwordChangedTime = Long.parseLong(passwordLastChangedTime);
-
-        int daysDifference = 0;
-        long currentTimeMillis = System.currentTimeMillis();
-        if (passwordChangedTime > 0) {
-            Calendar currentTime = Calendar.getInstance();
-            currentTime.add(Calendar.DATE, (int) currentTime.getTimeInMillis());
-            daysDifference = (int) ((currentTimeMillis - passwordChangedTime) / (1000 * 60 * 60 * 24));
-        }
-
-        int passwordExpiryInDays = PasswordPolicyConstants.PASSWORD_EXPIRY_IN_DAYS_DEFAULT_VALUE;
-
-        // Getting the configured number of days before password expiry in days
-        String passwordExpiryInDaysConfiguredValue = PasswordPolicyUtils.getResidentIdpProperty(tenantDomain,
-                PasswordPolicyConstants.PASSWORD_EXPIRY_IN_DAYS_FROM_CONFIG)
-                .orElseGet(() -> PasswordPolicyUtils.getIdentityEventProperty(
-                        PasswordPolicyConstants.PASSWORD_EXPIRY_IN_DAYS_FROM_CONFIG));
-
-        if (StringUtils.isNotBlank(passwordExpiryInDaysConfiguredValue)) {
-            passwordExpiryInDays = Integer.parseInt(passwordExpiryInDaysConfiguredValue);
-        }
-
-        return daysDifference > passwordExpiryInDays;
-    }
-
-    private static Optional<String> getClaimValue(org.wso2.carbon.user.api.UserStoreManager userStoreManager,
-                                                  String claimURI,
-                                                  String tenantAwareUsername) throws UserStoreException {
-
-        String[] claimURIs = new String[]{claimURI};
-        Map<String, String> claimValueMap =
-                userStoreManager.getUserClaimValues(tenantAwareUsername, claimURIs, null);
-        if (claimValueMap != null && !claimValueMap.isEmpty()) {
-            return Optional.of(claimValueMap.get(claimURI));
-        }
-        return Optional.empty();
-    }
-
-    private static String convertCreatedDateToEpochString(String createdDate) throws ParseException {
-
-        SimpleDateFormat simpleDateFormat = new SimpleDateFormat(PasswordPolicyConstants.CREATED_CLAIM_DATE_FORMAT);
-        simpleDateFormat.setTimeZone(TimeZone.getTimeZone(PasswordPolicyConstants.CREATED_CLAIM_TIMEZONE));
-
-        return String.valueOf(simpleDateFormat.parse(createdDate).getTime());
-    }
 }
